gstreamer1-plugins-good-1.0.10-1.fc19.stbtester1.src.rpm
Patched gstreamer1-plugins-good for Fedora 19

With the patches for https://bugzilla.gnome.org/show_bug.cgi?id=725860
to support the HDPVR with GStreamer 1.x.

These patches have been merged upstream and should be available in
gstreamer1-plugins-good 1.2.4 or 1.2.5 in the near future. In the
meantime you can use this package, published at
http://copr.fedoraproject.org/coprs/stbt/stb-tester/